Separated common SCC utilities into SCCBase transformation
